+++ plugins/processors/awsapplicationsignals/internal/resolver/kubernetes_test.go
@@ -9,15 +9,11 @@ import (
 	"strings"
 	"sync"
 	"testing"
-	"time"
 
-	mapset "github.com/deckarep/golang-set/v2"
 	"github.com/stretchr/testify/assert"
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	semconv "go.opentelemetry.io/collector/semconv/v1.22.0"
 	"go.uber.org/zap"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/aws/amazon-cloudwatch-agent/plugins/processors/awsapplicationsignals/common"
 	"github.com/aws/amazon-cloudwatch-agent/plugins/processors/awsapplicationsignals/config"
@@ -33,677 +29,6 @@ func (md *MockDeleter) DeleteWithDelay(m *sync.Map, key interface{}) {
 }
 
 var mockDeleter = &MockDeleter{}
-
-// TestAttachNamespace function
-func TestAttachNamespace(t *testing.T) {
-	result := attachNamespace("testResource", "testNamespace")
-	if result != "testResource@testNamespace" {
-		t.Errorf("attachNamespace was incorrect, got: %s, want: %s.", result, "testResource@testNamespace")
-	}
-}
-
-// TestGetServiceAndNamespace function
-func TestGetServiceAndNamespace(t *testing.T) {
-	service := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "testService",
-			Namespace: "testNamespace",
-		},
-	}
-	result := getServiceAndNamespace(service)
-	if result != "testService@testNamespace" {
-		t.Errorf("getServiceAndNamespace was incorrect, got: %s, want: %s.", result, "testService@testNamespace")
-	}
-}
-
-// TestExtractResourceAndNamespace function
-func TestExtractResourceAndNamespace(t *testing.T) {
-	// Test normal case
-	name, namespace := extractResourceAndNamespace("testService@testNamespace")
-	if name != "testService" || namespace != "testNamespace" {
-		t.Errorf("extractResourceAndNamespace was incorrect, got: %s and %s, want: %s and %s.", name, namespace, "testService", "testNamespace")
-	}
-
-	// Test invalid case
-	name, namespace = extractResourceAndNamespace("invalid")
-	if name != "" || namespace != "" {
-		t.Errorf("extractResourceAndNamespace was incorrect, got: %s and %s, want: %s and %s.", name, namespace, "", "")
-	}
-}
-
-func TestExtractWorkloadNameFromRS(t *testing.T) {
-	testCases := []struct {
-		name           string
-		replicaSetName string
-		want           string
-		shouldErr      bool
-	}{
-		{
-			name:           "Valid ReplicaSet Name",
-			replicaSetName: "my-deployment-5859ffc7ff",
-			want:           "my-deployment",
-			shouldErr:      false,
-		},
-		{
-			name:           "Invalid ReplicaSet Name - No Hyphen",
-			replicaSetName: "mydeployment5859ffc7ff",
-			want:           "",
-			shouldErr:      true,
-		},
-		{
-			name:           "Invalid ReplicaSet Name - Less Than 10 Suffix Characters",
-			replicaSetName: "my-deployment-bc2",
-			want:           "",
-			shouldErr:      true,
-		},
-		{
-			name:           "Invalid ReplicaSet Name - More Than 10 Suffix Characters",
-			replicaSetName: "my-deployment-5859ffc7ffx",
-			want:           "",
-			shouldErr:      true,
-		},
-		{
-			name:           "Invalid ReplicaSet Name - Invalid Characters in Suffix",
-			replicaSetName: "my-deployment-aeiou12345",
-			want:           "",
-			shouldErr:      true,
-		},
-		{
-			name:           "Invalid ReplicaSet Name - Empty String",
-			replicaSetName: "",
-			want:           "",
-			shouldErr:      true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got, err := extractWorkloadNameFromRS(tc.replicaSetName)
-
-			if (err != nil) != tc.shouldErr {
-				t.Errorf("extractWorkloadNameFromRS() error = %v, wantErr %v", err, tc.shouldErr)
-				return
-			}
-
-			if got != tc.want {
-				t.Errorf("extractWorkloadNameFromRS() = %v, want %v", got, tc.want)
-			}
-		})
-	}
-}
-
-func TestExtractWorkloadNameFromPodName(t *testing.T) {
-	testCases := []struct {
-		name      string
-		podName   string
-		want      string
-		shouldErr bool
-	}{
-		{
-			name:      "Valid Pod Name",
-			podName:   "my-replicaset-bc24f",
-			want:      "my-replicaset",
-			shouldErr: false,
-		},
-		{
-			name:      "Invalid Pod Name - No Hyphen",
-			podName:   "myreplicasetbc24f",
-			want:      "",
-			shouldErr: true,
-		},
-		{
-			name:      "Invalid Pod Name - Less Than 5 Suffix Characters",
-			podName:   "my-replicaset-bc2",
-			want:      "",
-			shouldErr: true,
-		},
-		{
-			name:      "Invalid Pod Name - More Than 5 Suffix Characters",
-			podName:   "my-replicaset-bc24f5",
-			want:      "",
-			shouldErr: true,
-		},
-		{
-			name:      "Invalid Pod Name - Empty String",
-			podName:   "",
-			want:      "",
-			shouldErr: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			got, err := extractWorkloadNameFromPodName(tc.podName)
-
-			if (err != nil) != tc.shouldErr {
-				t.Errorf("extractWorkloadNameFromPodName() error = %v, wantErr %v", err, tc.shouldErr)
-				return
-			}
-
-			if got != tc.want {
-				t.Errorf("extractWorkloadNameFromPodName() = %v, want %v", got, tc.want)
-			}
-		})
-	}
-}
-
-// TestGetWorkloadAndNamespace function
-func TestGetWorkloadAndNamespace(t *testing.T) {
-	// Test ReplicaSet case
-	pod := &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "testPod",
-			Namespace: "testNamespace",
-			OwnerReferences: []metav1.OwnerReference{
-				{
-					Kind: "ReplicaSet",
-					Name: "testDeployment-5d68bc5f49",
-				},
-			},
-		},
-	}
-	result := getWorkloadAndNamespace(pod)
-	if result != "testDeployment@testNamespace" {
-		t.Errorf("getDeploymentAndNamespace was incorrect, got: %s, want: %s.", result, "testDeployment@testNamespace")
-	}
-
-	// Test StatefulSet case
-	pod.ObjectMeta.OwnerReferences[0].Kind = "StatefulSet"
-	pod.ObjectMeta.OwnerReferences[0].Name = "testStatefulSet"
-	result = getWorkloadAndNamespace(pod)
-	if result != "testStatefulSet@testNamespace" {
-		t.Errorf("getWorkloadAndNamespace was incorrect, got: %s, want: %s.", result, "testStatefulSet@testNamespace")
-	}
-
-	// Test Other case
-	pod.ObjectMeta.OwnerReferences[0].Kind = "Other"
-	pod.ObjectMeta.OwnerReferences[0].Name = "testOther"
-	result = getWorkloadAndNamespace(pod)
-	if result != "" {
-		t.Errorf("getWorkloadAndNamespace was incorrect, got: %s, want: %s.", result, "")
-	}
-
-	// Test no OwnerReferences case
-	pod.ObjectMeta.OwnerReferences = nil
-	result = getWorkloadAndNamespace(pod)
-	if result != "" {
-		t.Errorf("getWorkloadAndNamespace was incorrect, got: %s, want: %s.", result, "")
-	}
-}
-
-func TestServiceToWorkloadMapper_MapServiceToWorkload(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
-
-	serviceAndNamespaceToSelectors := &sync.Map{}
-	workloadAndNamespaceToLabels := &sync.Map{}
-	serviceToWorkload := &sync.Map{}
-
-	serviceAndNamespaceToSelectors.Store("service1@namespace1", mapset.NewSet("label1=value1", "label2=value2"))
-	workloadAndNamespaceToLabels.Store("deployment1@namespace1", mapset.NewSet("label1=value1", "label2=value2", "label3=value3"))
-
-	mapper := newServiceToWorkloadMapper(serviceAndNamespaceToSelectors, workloadAndNamespaceToLabels, serviceToWorkload, logger, mockDeleter)
-	mapper.mapServiceToWorkload()
-
-	if _, ok := serviceToWorkload.Load("service1@namespace1"); !ok {
-		t.Errorf("Expected service1@namespace1 to be mapped to a workload, but it was not")
-	}
-}
-
-func TestServiceToWorkloadMapper_MapServiceToWorkload_NoWorkload(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
-
-	serviceAndNamespaceToSelectors := &sync.Map{}
-	workloadAndNamespaceToLabels := &sync.Map{}
-	serviceToWorkload := &sync.Map{}
-
-	// Add a service with no matching workload
-	serviceAndNamespace := "service@namespace"
-	serviceAndNamespaceToSelectors.Store(serviceAndNamespace, mapset.NewSet("label1=value1"))
-	serviceToWorkload.Store(serviceAndNamespace, "workload@namespace")
-
-	mapper := newServiceToWorkloadMapper(serviceAndNamespaceToSelectors, workloadAndNamespaceToLabels, serviceToWorkload, logger, mockDeleter)
-	mapper.mapServiceToWorkload()
-
-	// Check that the service was deleted from serviceToWorkload
-	if _, ok := serviceToWorkload.Load(serviceAndNamespace); ok {
-		t.Errorf("Service was not deleted from serviceToWorkload")
-	}
-}
-
-func TestServiceToWorkloadMapper_MapServiceToWorkload_MultipleWorkloads(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
-
-	serviceAndNamespaceToSelectors := &sync.Map{}
-	workloadAndNamespaceToLabels := &sync.Map{}
-	serviceToWorkload := &sync.Map{}
-
-	serviceAndNamespace := "service@namespace"
-	serviceAndNamespaceToSelectors.Store(serviceAndNamespace, mapset.NewSet("label1=value1", "label2=value2"))
-
-	// Add two workloads with matching labels to the service
-	workloadAndNamespaceToLabels.Store("workload1@namespace", mapset.NewSet("label1=value1", "label2=value2", "label3=value3"))
-	workloadAndNamespaceToLabels.Store("workload2@namespace", mapset.NewSet("label1=value1", "label2=value2", "label4=value4"))
-
-	mapper := newServiceToWorkloadMapper(serviceAndNamespaceToSelectors, workloadAndNamespaceToLabels, serviceToWorkload, logger, mockDeleter)
-	mapper.mapServiceToWorkload()
-
-	// Check that the service does not map to any workload
-	if _, ok := serviceToWorkload.Load(serviceAndNamespace); ok {
-		t.Errorf("Unexpected mapping of service to multiple workloads")
-	}
-}
-
-func TestMapServiceToWorkload_StopsWhenSignaled(t *testing.T) {
-	logger, _ := zap.NewDevelopment()
-
-	serviceAndNamespaceToSelectors := &sync.Map{}
-	workloadAndNamespaceToLabels := &sync.Map{}
-	serviceToWorkload := &sync.Map{}
-
-	stopchan := make(chan struct{})
-
-	// Signal the stopchan to stop after 100 milliseconds
-	time.AfterFunc(100*time.Millisecond, func() {
-		close(stopchan)
-	})
-
-	mapper := newServiceToWorkloadMapper(serviceAndNamespaceToSelectors, workloadAndNamespaceToLabels, serviceToWorkload, logger, mockDeleter)
-
-	start := time.Now()
-	mapper.Start(stopchan)
-	duration := time.Since(start)
-
-	// Check that the function stopped in a reasonable time after the stop signal
-	if duration > 200*time.Millisecond {
-		t.Errorf("mapServiceToWorkload did not stop in a reasonable time after the stop signal, duration: %v", duration)
-	}
-}
-
-func TestOnAddOrUpdateService(t *testing.T) {
-	// Create a fake service
-	service := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myservice",
-			Namespace: "mynamespace",
-		},
-		Spec: corev1.ServiceSpec{
-			ClusterIP: "1.2.3.4",
-			Selector: map[string]string{
-				"app": "myapp",
-			},
-		},
-	}
-
-	// Create the maps
-	ipToServiceAndNamespace := &sync.Map{}
-	serviceAndNamespaceToSelectors := &sync.Map{}
-
-	// Call the function
-	svcWatcher := newServiceWatcherForTesting(ipToServiceAndNamespace, serviceAndNamespaceToSelectors)
-	svcWatcher.onAddOrUpdateService(service)
-
-	// Check that the maps contain the expected entries
-	if _, ok := ipToServiceAndNamespace.Load("1.2.3.4"); !ok {
-		t.Errorf("ipToServiceAndNamespace does not contain the service IP")
-	}
-	if _, ok := serviceAndNamespaceToSelectors.Load("myservice@mynamespace"); !ok {
-		t.Errorf("serviceAndNamespaceToSelectors does not contain the service")
-	}
-}
-
-func TestOnDeleteService(t *testing.T) {
-	// Create a fake service
-	service := &corev1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "myservice",
-			Namespace: "mynamespace",
-		},
-		Spec: corev1.ServiceSpec{
-			ClusterIP: "1.2.3.4",
-			Selector: map[string]string{
-				"app": "myapp",
-			},
-		},
-	}
-
-	// Create the maps and add the service to them
-	ipToServiceAndNamespace := &sync.Map{}
-	ipToServiceAndNamespace.Store("1.2.3.4", "myservice@mynamespace")
-	serviceAndNamespaceToSelectors := &sync.Map{}
-	serviceAndNamespaceToSelectors.Store("myservice@mynamespace", mapset.NewSet("app=myapp"))
-
-	// Call the function
-	svcWatcher := newServiceWatcherForTesting(ipToServiceAndNamespace, serviceAndNamespaceToSelectors)
-	svcWatcher.onDeleteService(service, mockDeleter)
-
-	// Check that the maps do not contain the service
-	if _, ok := ipToServiceAndNamespace.Load("1.2.3.4"); ok {
-		t.Errorf("ipToServiceAndNamespace still contains the service IP")
-	}
-	if _, ok := serviceAndNamespaceToSelectors.Load("myservice@mynamespace"); ok {
-		t.Errorf("serviceAndNamespaceToSelectors still contains the service")
-	}
-}
-
-func TestOnAddOrUpdatePod(t *testing.T) {
-	t.Run("pod with both PodIP and HostIP", func(t *testing.T) {
-		ipToPod := &sync.Map{}
-		podToWorkloadAndNamespace := &sync.Map{}
-		workloadAndNamespaceToLabels := &sync.Map{}
-		workloadPodCount := map[string]int{}
-
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "testPod",
-				Namespace: "testNamespace",
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Kind: "ReplicaSet",
-						Name: "testDeployment-598b89cd8d",
-					},
-				},
-			},
-			Status: corev1.PodStatus{
-				PodIP:  "1.2.3.4",
-				HostIP: "5.6.7.8",
-			},
-		}
-
-		poWatcher := newPodWatcherForTesting(ipToPod, podToWorkloadAndNamespace, workloadAndNamespaceToLabels, workloadPodCount)
-		poWatcher.onAddOrUpdatePod(pod, nil)
-
-		// Test the mappings in ipToPod
-		if podName, _ := ipToPod.Load("1.2.3.4"); podName != "testPod" {
-			t.Errorf("ipToPod was incorrect, got: %s, want: %s.", podName, "testPod")
-		}
-
-		// Test the mapping in podToWorkloadAndNamespace
-		if depAndNamespace, _ := podToWorkloadAndNamespace.Load("testPod"); depAndNamespace != "testDeployment@testNamespace" {
-			t.Errorf("podToWorkloadAndNamespace was incorrect, got: %s, want: %s.", depAndNamespace, "testDeployment@testNamespace")
-		}
-
-		// Test the count in workloadPodCount
-		if count := workloadPodCount["testDeployment@testNamespace"]; count != 1 {
-			t.Errorf("workloadPodCount was incorrect, got: %d, want: %d.", count, 1)
-		}
-	})
-
-	t.Run("pod with only HostIP", func(t *testing.T) {
-		ipToPod := &sync.Map{}
-		podToWorkloadAndNamespace := &sync.Map{}
-		workloadAndNamespaceToLabels := &sync.Map{}
-		workloadPodCount := map[string]int{}
-
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "testPod",
-				Namespace: "testNamespace",
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Kind: "ReplicaSet",
-						Name: "testDeployment-7b74958fb8",
-					},
-				},
-			},
-			Status: corev1.PodStatus{
-				HostIP: "5.6.7.8",
-			},
-			Spec: corev1.PodSpec{
-				HostNetwork: true,
-				Containers: []corev1.Container{
-					{
-						Ports: []corev1.ContainerPort{
-							{
-								HostPort: int32(8080),
-							},
-						},
-					},
-				},
-			},
-		}
-
-		poWatcher := newPodWatcherForTesting(ipToPod, podToWorkloadAndNamespace, workloadAndNamespaceToLabels, workloadPodCount)
-		poWatcher.onAddOrUpdatePod(pod, nil)
-
-		// Test the mappings in ipToPod
-		if podName, _ := ipToPod.Load("5.6.7.8:8080"); podName != "testPod" {
-			t.Errorf("ipToPod was incorrect, got: %s, want: %s.", podName, "testPod")
-		}
-
-		// Test the mapping in podToWorkloadAndNamespace
-		if depAndNamespace, _ := podToWorkloadAndNamespace.Load("testPod"); depAndNamespace != "testDeployment@testNamespace" {
-			t.Errorf("podToWorkloadAndNamespace was incorrect, got: %s, want: %s.", depAndNamespace, "testDeployment@testNamespace")
-		}
-
-		// Test the count in workloadPodCount
-		if count := workloadPodCount["testDeployment@testNamespace"]; count != 1 {
-			t.Errorf("workloadPodCount was incorrect, got: %d, want: %d.", count, 1)
-		}
-	})
-
-	t.Run("pod updated with different set of labels", func(t *testing.T) {
-		ipToPod := &sync.Map{}
-		podToWorkloadAndNamespace := &sync.Map{}
-		workloadAndNamespaceToLabels := &sync.Map{}
-
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "testPod",
-				Namespace: "testNamespace",
-				Labels: map[string]string{
-					"label1": "value1",
-					"label2": "value2",
-				},
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Kind: "ReplicaSet",
-						Name: "testDeployment-5d68bc5f49",
-					},
-				},
-			},
-			Status: corev1.PodStatus{
-				HostIP: "5.6.7.8",
-			},
-			Spec: corev1.PodSpec{
-				HostNetwork: true,
-				Containers: []corev1.Container{
-					{
-						Ports: []corev1.ContainerPort{
-							{HostPort: 8080},
-						},
-					},
-				},
-			},
-		}
-
-		// add the pod
-		poWatcher := newPodWatcherForTesting(ipToPod, podToWorkloadAndNamespace, workloadAndNamespaceToLabels, map[string]int{})
-		poWatcher.onAddOrUpdatePod(pod, nil)
-
-		// Test the mappings in ipToPod
-		if podName, ok := ipToPod.Load("5.6.7.8:8080"); !ok && podName != "testPod" {
-			t.Errorf("ipToPod[%s] was incorrect, got: %s, want: %s.", "5.6.7.8:8080", podName, "testPod")
-		}
-
-		// Test the mapping in workloadAndNamespaceToLabels
-		labels, _ := workloadAndNamespaceToLabels.Load("testDeployment@testNamespace")
-		expectedLabels := []string{"label1=value1", "label2=value2"}
-		for _, label := range expectedLabels {
-			if !labels.(mapset.Set[string]).Contains(label) {
-				t.Errorf("deploymentAndNamespaceToLabels was incorrect, got: %v, want: %s.", labels, label)
-			}
-		}
-
-		pod2 := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "testPod",
-				Namespace: "testNamespace",
-				Labels: map[string]string{
-					"label1": "value1",
-					"label2": "value2",
-					"label3": "value3",
-				},
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Kind: "ReplicaSet",
-						Name: "testDeployment-5d68bc5f49",
-					},
-				},
-			},
-			Status: corev1.PodStatus{
-				PodIP:  "1.2.3.4",
-				HostIP: "5.6.7.8",
-			},
-			Spec: corev1.PodSpec{
-				HostNetwork: true,
-				Containers: []corev1.Container{
-					{
-						Ports: []corev1.ContainerPort{
-							{HostPort: 8080},
-						},
-					},
-				},
-			},
-		}
-
-		// add the pod
-		poWatcher.onAddOrUpdatePod(pod2, pod)
-
-		// Test the mappings in ipToPod
-		if podName, ok := ipToPod.Load("5.6.7.8:8080"); !ok && podName != "testPod" {
-			t.Errorf("ipToPod[%s] was incorrect, got: %s, want: %s.", "5.6.7.8:8080", podName, "testPod")
-		}
-
-		if podName, ok := ipToPod.Load("1.2.3.4"); !ok && podName != "testPod" {
-			t.Errorf("ipToPod[%s] was incorrect, got: %s, want: %s.", "1.2.3.4", podName, "testPod")
-		}
-		// Test the mapping in workloadAndNamespaceToLabels
-		labels, _ = workloadAndNamespaceToLabels.Load("testDeployment@testNamespace")
-		expectedLabels = []string{"label1=value1", "label2=value2", "label3=value3"}
-		for _, label := range expectedLabels {
-			if !labels.(mapset.Set[string]).Contains(label) {
-				t.Errorf("workloadAndNamespaceToLabels was incorrect, got: %v, want: %s.", labels, label)
-			}
-		}
-	})
-}
-
-func TestOnDeletePod(t *testing.T) {
-	t.Run("pod with both PodIP and HostIP", func(t *testing.T) {
-		ipToPod := &sync.Map{}
-		podToWorkloadAndNamespace := &sync.Map{}
-		workloadAndNamespaceToLabels := &sync.Map{}
-		workloadPodCount := map[string]int{}
-
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "testPod",
-				Namespace: "testNamespace",
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Kind: "ReplicaSet",
-						Name: "testDeployment-xyz",
-					},
-				},
-			},
-			Status: corev1.PodStatus{
-				PodIP:  "1.2.3.4",
-				HostIP: "5.6.7.8",
-			},
-		}
-
-		// Assume the pod has already been added
-		ipToPod.Store(pod.Status.PodIP, pod.Name)
-		ipToPod.Store(pod.Status.HostIP, pod.Name)
-		podToWorkloadAndNamespace.Store(pod.Name, "testDeployment@testNamespace")
-		workloadAndNamespaceToLabels.Store("testDeployment@testNamespace", "testLabels")
-		workloadPodCount["testDeployment@testNamespace"] = 1
-
-		poWatcher := newPodWatcherForTesting(ipToPod, podToWorkloadAndNamespace, workloadAndNamespaceToLabels, workloadPodCount)
-		poWatcher.onDeletePod(pod)
-
-		// Test if the entries in ipToPod and podToWorkloadAndNamespace have been deleted
-		if _, ok := ipToPod.Load("1.2.3.4"); ok {
-			t.Errorf("ipToPod deletion was incorrect, key: %s still exists", "1.2.3.4")
-		}
-
-		if _, ok := podToWorkloadAndNamespace.Load("testPod"); ok {
-			t.Errorf("podToWorkloadAndNamespace deletion was incorrect, key: %s still exists", "testPod")
-		}
-
-		// Test if the count in workloadPodCount has been decremented and the entry in workloadAndNamespaceToLabels has been deleted
-		if count := workloadPodCount["testDeployment@testNamespace"]; count != 0 {
-			t.Errorf("workloadPodCount was incorrect, got: %d, want: %d.", count, 0)
-		}
-
-		if _, ok := workloadAndNamespaceToLabels.Load("testDeployment@testNamespace"); ok {
-			t.Errorf("workloadAndNamespaceToLabels deletion was incorrect, key: %s still exists", "testDeployment@testNamespace")
-		}
-	})
-
-	t.Run("pod with only HostIP and some network ports", func(t *testing.T) {
-		ipToPod := &sync.Map{}
-		podToWorkloadAndNamespace := &sync.Map{}
-		workloadAndNamespaceToLabels := &sync.Map{}
-		workloadPodCount := map[string]int{}
-
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "testPod",
-				Namespace: "testNamespace",
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						Kind: "ReplicaSet",
-						Name: "testDeployment-xyz",
-					},
-				},
-			},
-			Status: corev1.PodStatus{
-				HostIP: "5.6.7.8",
-			},
-			Spec: corev1.PodSpec{
-				HostNetwork: true,
-				Containers: []corev1.Container{
-					{
-						Ports: []corev1.ContainerPort{
-							{
-								HostPort: int32(8080),
-							},
-						},
-					},
-				},
-			},
-		}
-
-		// Assume the pod has already been added
-		ipToPod.Store(pod.Status.HostIP, pod.Name)
-		ipToPod.Store(pod.Status.HostIP+":8080", pod.Name)
-		podToWorkloadAndNamespace.Store(pod.Name, "testDeployment@testNamespace")
-		workloadAndNamespaceToLabels.Store("testDeployment@testNamespace", "testLabels")
-		workloadPodCount["testDeployment@testNamespace"] = 1
-
-		poWatcher := newPodWatcherForTesting(ipToPod, podToWorkloadAndNamespace, workloadAndNamespaceToLabels, workloadPodCount)
-		poWatcher.onDeletePod(pod)
-
-		// Test if the entries in ipToPod and podToWorkloadAndNamespace have been deleted
-		if _, ok := ipToPod.Load("5.6.7.8:8080"); ok {
-			t.Errorf("ipToPod deletion was incorrect, key: %s still exists", "5.6.7.8:8080")
-		}
-
-		if _, ok := podToWorkloadAndNamespace.Load("testPod"); ok {
-			t.Errorf("podToDeploymentAndNamespace deletion was incorrect, key: %s still exists", "testPod")
-		}
-
-		// Test if the count in workloadPodCount has been decremented and the entry in workloadAndNamespaceToLabels has been deleted
-		if count := workloadPodCount["testDeployment@testNamespace"]; count != 0 {
-			t.Errorf("workloadPodCount was incorrect, got: %d, want: %d.", count, 0)
-		}
-
-		if _, ok := workloadAndNamespaceToLabels.Load("testDeployment@testNamespace"); ok {
-			t.Errorf("workloadAndNamespaceToLabels deletion was incorrect, key: %s still exists", "testDeployment@testNamespace")
-		}
-	})
-}
 
 func TestEksResolver(t *testing.T) {
 	logger, _ := zap.NewProduction()
@@ -1062,232 +387,5 @@ func TestK8sResourceAttributesResolverOnK8SOnPrem(t *testing.T) {
 			_, exists = attributes.Get(common.AttributeEC2InstanceId)
 			assert.False(t, exists)
 		})
-	}
-}
-
-func TestExtractIPPort(t *testing.T) {
-	// Test valid IP:Port
-	ip, port, ok := extractIPPort("192.0.2.0:8080")
-	assert.Equal(t, "192.0.2.0", ip)
-	assert.Equal(t, "8080", port)
-	assert.True(t, ok)
-
-	// Test invalid IP:Port
-	ip, port, ok = extractIPPort("192.0.2:8080")
-	assert.Equal(t, "", ip)
-	assert.Equal(t, "", port)
-	assert.False(t, ok)
-
-	// Test IP only
-	ip, port, ok = extractIPPort("192.0.2.0")
-	assert.Equal(t, "", ip)
-	assert.Equal(t, "", port)
-	assert.False(t, ok)
-}
-
-func TestFilterPodIPFields(t *testing.T) {
-	meta := metav1.ObjectMeta{
-		Name:      "test",
-		Namespace: "default",
-		Labels: map[string]string{
-			"name": "app",
-		},
-	}
-	pod := &corev1.Pod{
-		ObjectMeta: meta,
-		Spec: corev1.PodSpec{
-			HostNetwork: true,
-			Containers: []corev1.Container{
-				{},
-			},
-		},
-		Status: corev1.PodStatus{},
-	}
-	newPod, err := minimizePod(pod)
-	assert.Nil(t, err)
-	assert.Empty(t, getHostNetworkPorts(newPod.(*corev1.Pod)))
-
-	podStatus := corev1.PodStatus{
-		PodIP: "192.168.0.12",
-		HostIPs: []corev1.HostIP{
-			{
-				IP: "132.168.3.12",
-			},
-		},
-	}
-	pod = &corev1.Pod{
-		ObjectMeta: meta,
-		Spec: corev1.PodSpec{
-			HostNetwork: true,
-			Containers: []corev1.Container{
-				{
-					Ports: []corev1.ContainerPort{
-						{HostPort: 8080},
-					},
-				},
-			},
-		},
-		Status: podStatus,
-	}
-	newPod, err = minimizePod(pod)
-	assert.Nil(t, err)
-	assert.Equal(t, "app", newPod.(*corev1.Pod).Labels["name"])
-	assert.Equal(t, []string{"8080"}, getHostNetworkPorts(newPod.(*corev1.Pod)))
-	assert.Equal(t, podStatus, newPod.(*corev1.Pod).Status)
-
-	pod = &corev1.Pod{
-		Spec: corev1.PodSpec{
-			HostNetwork: true,
-			Containers: []corev1.Container{
-				{
-					Ports: []corev1.ContainerPort{
-						{HostPort: 8080},
-						{HostPort: 8081},
-					},
-				},
-			},
-		},
-		Status: podStatus,
-	}
-	newPod, err = minimizePod(pod)
-	assert.Nil(t, err)
-	assert.Equal(t, []string{"8080", "8081"}, getHostNetworkPorts(newPod.(*corev1.Pod)))
-	assert.Equal(t, podStatus, newPod.(*corev1.Pod).Status)
-}
-
-func TestFilterServiceIPFields(t *testing.T) {
-	meta := metav1.ObjectMeta{
-		Name:      "test",
-		Namespace: "default",
-	}
-	svc := &corev1.Service{
-		ObjectMeta: meta,
-		Spec: corev1.ServiceSpec{
-			Selector: map[string]string{
-				"name": "app",
-			},
-			ClusterIP: "10.0.12.4",
-		},
-	}
-	newSvc, err := minimizeService(svc)
-	assert.Nil(t, err)
-	assert.Equal(t, "10.0.12.4", newSvc.(*corev1.Service).Spec.ClusterIP)
-	assert.Equal(t, "app", newSvc.(*corev1.Service).Spec.Selector["name"])
-}
-
-func TestHandlePodUpdate(t *testing.T) {
-	testCases := []struct {
-		name            string
-		oldPod          *corev1.Pod
-		newPod          *corev1.Pod
-		initialIPToPod  map[string]string
-		expectedIPToPod map[string]string
-	}{
-		{
-			name: "Old and New Pod Do Not Use Host Network, Different Pod IPs",
-			oldPod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "mypod",
-				},
-				Status: corev1.PodStatus{
-					PodIP: "10.0.0.3",
-				},
-				Spec: corev1.PodSpec{
-					HostNetwork: false,
-				},
-			},
-			newPod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "mypod",
-				},
-				Status: corev1.PodStatus{
-					PodIP: "10.0.0.4",
-				},
-				Spec: corev1.PodSpec{
-					HostNetwork: false,
-				},
-			},
-			initialIPToPod: map[string]string{
-				"10.0.0.3": "mypod",
-			},
-			expectedIPToPod: map[string]string{
-				"10.0.0.4": "mypod",
-			},
-		},
-		{
-			name: "Old Pod Has Empty PodIP, New Pod Does Not Use Host Network, Non-Empty Pod IP",
-			oldPod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "mypod",
-				},
-				Status: corev1.PodStatus{
-					PodIP: "",
-				},
-				Spec: corev1.PodSpec{
-					HostNetwork: false,
-				},
-			},
-			newPod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "mypod",
-				},
-				Status: corev1.PodStatus{
-					PodIP: "10.0.0.5",
-				},
-				Spec: corev1.PodSpec{
-					HostNetwork: false,
-				},
-			},
-			initialIPToPod: map[string]string{},
-			expectedIPToPod: map[string]string{
-				"10.0.0.5": "mypod",
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			ipToPod := &sync.Map{}
-			// Initialize ipToPod map
-			for k, v := range tc.initialIPToPod {
-				ipToPod.Store(k, v)
-			}
-			poWatcher := newPodWatcherForTesting(ipToPod, nil, nil, map[string]int{})
-			poWatcher.handlePodUpdate(tc.newPod, tc.oldPod)
-
-			// Now validate that ipToPod map has been updated correctly
-			for key, expectedValue := range tc.expectedIPToPod {
-				val, ok := ipToPod.Load(key)
-				if !ok || val.(string) != expectedValue {
-					t.Errorf("Expected record for %v to be %v, got %v", key, expectedValue, val)
-				}
-			}
-			// Validate that old keys have been removed
-			for key := range tc.initialIPToPod {
-				if _, ok := tc.expectedIPToPod[key]; !ok {
-					if _, ok := ipToPod.Load(key); ok {
-						t.Errorf("Expected record for %v to be removed, but it was not", key)
-					}
-				}
-			}
-		})
-	}
-}
-
-func newServiceWatcherForTesting(ipToServiceAndNamespace, serviceAndNamespaceToSelectors *sync.Map) *serviceWatcher {
-	logger, _ := zap.NewDevelopment()
-	return &serviceWatcher{ipToServiceAndNamespace, serviceAndNamespaceToSelectors, logger, nil, nil}
-}
-
-func newPodWatcherForTesting(ipToPod, podToWorkloadAndNamespace, workloadAndNamespaceToLabels *sync.Map, workloadPodCount map[string]int) *podWatcher {
-	logger, _ := zap.NewDevelopment()
-	return &podWatcher{
-		ipToPod:                      ipToPod,
-		podToWorkloadAndNamespace:    podToWorkloadAndNamespace,
-		workloadAndNamespaceToLabels: workloadAndNamespaceToLabels,
-		workloadPodCount:             workloadPodCount,
-		logger:                       logger,
-		informer:                     nil,
-		deleter:                      mockDeleter,
 	}
 }
